Completely implemented TaskSerializableInteractorDataIn and TaskSerializableInteractorDataOut and fixed all bugs resulting from the chanfes in UserInterface.
