--- src/main/java/com/csc207/api/WeekSerializableDataIn.java
@@ -1,4 +1,0 @@
-package com.csc207.api;
-
-public class WeekSerializableDataIn {
-}
