+++ src/main/java/com/csc207/api/WeekSerializableInteractorDataIn.java
@@ -1,0 +1,4 @@
+package com.csc207.api;
+
+public class WeekSerializableInteractorDataIn {
+}
